Filesystem Encoding Methods

+++ src/Stash/Driver/FileSystem.php
@@ -12,6 +12,8 @@
 namespace Stash\Driver;
 
 use Stash;
+use Stash\Driver\FileSystem\NativeEncoder;
+use Stash\Driver\FileSystem\EncoderInterface;
 use Stash\Utilities;
 use Stash\Exception\LogicException;
 use Stash\Exception\RuntimeException;
@@ -115,6 +117,11 @@ class FileSystem implements DriverInterface
     );
 
     /**
+     * @var \Stash\Driver\FileSystem\EncoderInterface
+     */
+    protected $encoder;
+
+    /**
      * Requests a list of options.
      *
      * @param  array                             $options
@@ -147,6 +154,25 @@ class FileSystem implements DriverInterface
         }
 
         $this->memStoreLimit = (int) $options['memKeyLimit'];
+
+        if (isset($options['encoder'])) {
+            $encoder = $options['encoder'];
+            if (is_object($encoder)) {
+                if (!($encoder instanceof EncoderInterface)) {
+                    throw new RuntimeException('Encoder object must implement EncoderInterface');
+                }
+                $this->encoder = new $encoder;
+            } else {
+                $encoderClass = 'Stash\Driver\FileSystem\\' . $encoder . 'Encoder';
+                if (class_exists($encoder)) {
+                    $this->encoder = new $encoder();
+                } elseif (class_exists($encoderClass)) {
+                    $this->encoder = new $encoderClass();
+                } else {
+                    throw new RuntimeException('Invalid Encoder: ' . $encoder);
+                }
+            }
+        }
 
         Utilities::checkFileSystemPermissions($this->cachePath, $this->dirPermissions);
     }
@@ -185,41 +211,7 @@ class FileSystem implements DriverInterface
      */
     public function getData($key)
     {
-        return self::getDataFromFile($this->makePath($key));
-    }
-
-    /**
-     * Retrieves data from file if that data is there.
-     *
-     * @param  string     $path
-     * @return array|bool
-     */
-    protected static function getDataFromFile($path)
-    {
-        if (!file_exists($path)) {
-            return false;
-        }
-
-        include($path);
-
-        if (!isset($loaded)) {
-            return false;
-        }
-
-        if (!isset($expiration)) {
-            $expiration = null;
-        }
-
-        // If the item does not exist we should return false. However, it's
-        // possible that the item exists as null, so we have to make sure that
-        // it's both unset and not null. The downside to this is that the
-        // is_null function will issue a warning on an item that isn't set.
-        // So we're stuck testing and suppressing the warning.
-        if (!isset($data) || @is_null($data)) {
-            return array('data' => null, 'expiration' => $expiration);
-        } else {
-            return array('data' => $data, 'expiration' => $expiration);
-        }
+        return $this->getEncoder()->deserialize($this->makePath($key));
     }
 
     /**
@@ -249,38 +241,7 @@ class FileSystem implements DriverInterface
             }
         }
 
-        $storeString = '<?php ' . PHP_EOL
-            . '/* Cachekey: ' . str_replace('*/', '', $this->makeKeyString($key)) . ' */' . PHP_EOL
-            . '/* Type: ' . gettype($data) . ' */' . PHP_EOL
-            . '/* Expiration: ' . (isset($expiration) ? date(DATE_W3C, $expiration) : 'none') . ' */' . PHP_EOL
-            . PHP_EOL
-            . PHP_EOL
-            . PHP_EOL
-            . '$loaded = true;' . PHP_EOL;
-
-        if (isset($expiration)) {
-            $storeString .= '$expiration = ' . $expiration . ';' . PHP_EOL;
-        }
-
-        $storeString .= PHP_EOL;
-
-        if (is_array($data)) {
-            $storeString .= "\$data = array();" . PHP_EOL;
-
-            foreach ($data as $key => $value) {
-                $dataString = $this->encode($value);
-                $keyString = "'" . str_replace("'", "\\'", $key) . "'";
-                $storeString .= PHP_EOL;
-                $storeString .= '/* Child Type: ' . gettype($value) . ' */' . PHP_EOL;
-                $storeString .= "\$data[{$keyString}] = {$dataString};" . PHP_EOL;
-            }
-        } else {
-
-            $dataString = $this->encode($data);
-            $storeString .= '/* Type: ' . gettype($data) . ' */' . PHP_EOL;
-            $storeString .= "\$data = {$dataString};" . PHP_EOL;
-        }
-
+        $storeString = $this->getEncoder()->serialize($this->makeKeyString($key), $data, $expiration);
         $result = file_put_contents($path, $storeString, LOCK_EX);
 
         // If opcache is switched on, it will try to cache the PHP data file
@@ -293,40 +254,6 @@ class FileSystem implements DriverInterface
         }
 
         return false !== $result;
-    }
-
-    /**
-     * Finds the method of encoding that has the cheapest decode needs and encodes the data with that method.
-     *
-     * @param  string $data
-     * @return string
-     */
-    protected function encode($data)
-    {
-        switch (Utilities::encoding($data)) {
-            case 'bool':
-                $dataString = (bool) $data ? 'true' : 'false';
-                break;
-
-            case 'serialize':
-                $dataString = 'unserialize(base64_decode(\'' . base64_encode(serialize($data)) . '\'))';
-                break;
-
-            case 'string':
-                $dataString = sprintf('"%s"', addcslashes($data, "\t\"\$\\"));
-                break;
-
-            case 'none':
-            default :
-                if (is_numeric($data)) {
-                    $dataString = (string) $data;
-                } else {
-                    $dataString = 'base64_decode(\'' . base64_encode($data) . '\')';
-                }
-                break;
-        }
-
-        return $dataString;
     }
 
     /**
@@ -384,7 +311,7 @@ class FileSystem implements DriverInterface
                 }
             }
 
-            $path = rtrim($path, DIRECTORY_SEPARATOR) . '.php';
+            $path = rtrim($path, DIRECTORY_SEPARATOR) . $this->getEncoder()->getExtension();
             $this->memStore['keys'][$memkey] = $path;
 
             // in most cases the key will be used almost immediately or not at all, so it doesn't need to grow too large
@@ -412,8 +339,9 @@ class FileSystem implements DriverInterface
             unlink($path);
         }
 
-        if (strpos($path, '.php') !== false) {
-            $path = substr($path, 0, -4);
+        $extension = $this->getEncoder()->getExtension();
+        if (strpos($path, $extension) !== false) {
+            $path = substr($path, 0, -(strlen($extension)));
         }
 
         if (is_dir($path)) {
@@ -451,8 +379,9 @@ class FileSystem implements DriverInterface
                 continue;
             }
 
-            $data = self::getDataFromFile($filename);
-            if ($data['expiration'] <= $startTime) {
+            $data = $this->getEncoder()->deserialize($filename);
+
+            if (is_numeric($data['expiration']) && $data['expiration'] <= $startTime) {
                 unlink($filename);
             }
 
@@ -460,6 +389,15 @@ class FileSystem implements DriverInterface
         unset($directoryIt);
 
         return true;
+    }
+
+    protected function getEncoder()
+    {
+        if (!isset($this->encoder)) {
+            $this->encoder = new \Stash\Driver\FileSystem\NativeEncoder();
+        }
+
+        return $this->encoder;
     }
 
     /**
